Do some (ULONG) casts when trying to make sure that new size is not bigger than MaxWidth, MaxHeight. Otherwise a MaxWidth, MaxHeight of -1 causes trouble.
git-svn-id: https://svn.aros.org/svn/aros/trunk@28488 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/intuition/zipwindow.c
@@ -113,8 +113,8 @@ static VOID int_zipwindow(struct ZipWindowActionMsg *msg,
         NewWidth    = w->ZipWidth;
 	if (window->Flags & WFLG_SIZEGADGET)
 	{
-            if (NewWidth < window->MinWidth) NewWidth = window->MinWidth;
-            if (NewWidth > window->MaxWidth) NewWidth = window->MaxWidth;
+            if (NewWidth < (ULONG)window->MinWidth) NewWidth = window->MinWidth;
+            if (NewWidth > (ULONG)window->MaxWidth) NewWidth = window->MaxWidth;
 	}
         w->ZipWidth = w->window.Width;
     }
@@ -125,8 +125,8 @@ static VOID int_zipwindow(struct ZipWindowActionMsg *msg,
         NewHeight    = w->ZipHeight;
 	if (window->Flags & WFLG_SIZEGADGET)
 	{
-            if (NewHeight < window->MinHeight) NewHeight = window->MinHeight;
-            if (NewHeight > window->MaxHeight) NewHeight = window->MaxHeight;
+            if (NewHeight < (ULONG)window->MinHeight) NewHeight = window->MinHeight;
+            if (NewHeight > (ULONG)window->MaxHeight) NewHeight = window->MaxHeight;
 	}
         w->ZipHeight = w->window.Height;
     }
